+++ Changelog.md
@@ -21,6 +21,7 @@ Ruby 2.1 is no longer officially supported.
 ## Bug fixes
 
 * Fix height too high on mobile SPV [#7480](https://github.com/diaspora/diaspora/pull/7480)
+* Improve stream when ignoring a person who posts a lot of tagged posts [#7503](https://github.com/diaspora/diaspora/pull/7503)
 
 ## Features
 * Add support for mentions in comments to the backend [#6818](https://github.com/diaspora/diaspora/pull/6818)